Refactor some stuff
