+++ src/interactive-window/serviceRegistry.web.ts
@@ -2,12 +2,9 @@
 // Licensed under the MIT License.
 'use strict';
 
-import { IExtensionSingleActivationService } from '../platform/activation/types';
 import { IServiceManager } from '../platform/ioc/types';
-import { CommandRegistry } from './commands/commandRegistry.web';
 import { CellHashProviderFactory } from './editor-integration/cellHashProviderFactory';
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<CellHashProviderFactory>(CellHashProviderFactory, CellHashProviderFactory);
-    serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, CommandRegistry);
 }
